refactor(ui/settings): terminal-aligned redesign
Bring the Settings page in line with the rest of the app's terminal
aesthetic. Major changes:

- Extract ThemeSection from SettingsPage.jsx into its own
  AppearanceSection.jsx so all four panels live in their own component.
- Replace each panel's <h2 className="settings-section-title"> with the
  shared SectionLabel terminal component (▶ ANALYSIS / ▶ SERVER /
  ▶ APPEARANCE / ▶ ABOUT).
- Introduce .settings-pill / .settings-pill--active / .settings-pill-group
  matching .term-flag-pill (mono, tight, subtle border, soft accent fill
  when active). Apply across provider tabs, model client selector, theme
  mode + family pickers, time-limit toggle, and advanced toggles.
- Drop the .settings-header / .settings-body / .settings-body--full
  wrapper divs from SettingsPage.jsx. Place TermHeader directly as a
  flex child of .settings-page so the page inherits padding from the
  parent .dashboard, matching map / violations / history.
- Make the grid full-width (no max-width cap), tighten panel-header,
  settings-row, and server-status padding.

Remove the now-unused .settings-section-title, .settings-header,
.settings-body, .theme-toggle*, .theme-family-*, and .provider-tab*
CSS rules.
