+++ src/quodeq/ui/src/styles/base.css
@@ -1051,55 +1051,11 @@ textarea:focus {
 /* ─── Settings page ─── */
 
 .settings-page {
-  width: 100%;
   display: flex;
   flex-direction: column;
-  min-height: 100%;
-}
-
-.settings-header {
-  padding: var(--space-8) var(--space-8) var(--space-6);
-  border-bottom: 1px solid var(--color-border);
-}
-
-.settings-page-icon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 40px;
-  height: 40px;
-  background: var(--color-accent-soft);
-  border-radius: var(--radius-lg);
-  color: var(--color-accent);
-  flex-shrink: 0;
-}
-
-.settings-title {
-  font-size: var(--text-lg);
-  font-weight: var(--weight-semibold);
-  color: var(--color-text);
-  margin: 0 0 2px 0;
-}
-
-.settings-subtitle {
-  margin: 0;
-  font-size: var(--text-sm);
-  color: var(--color-text-muted);
-}
-
-.settings-layout {
-  flex: 1;
-  min-height: 0;
-}
-
-.settings-body {
-  padding: var(--space-8);
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-5);
+  gap: var(--space-4);
+  min-width: 0;
   width: 100%;
-  max-width: 760px;
-  flex-shrink: 0;
 }
 
 .brand-carousel {
@@ -1108,15 +1064,12 @@ textarea:focus {
   gap: var(--space-4);
 }
 
-.settings-body--full {
-  max-width: 100%;
-}
-
 .settings-grid {
   display: grid;
   grid-template-columns: 1fr 1fr;
   gap: var(--space-5);
   align-items: start;
+  width: 100%;
 }
 
 .settings-grid-col {
@@ -1128,18 +1081,6 @@ textarea:focus {
 @media (max-width: 900px) {
   .settings-grid {
     grid-template-columns: 1fr;
-  }
-
-  .settings-header {
-    padding: var(--space-5) var(--space-5) var(--space-4);
-  }
-
-  .settings-body {
-    padding: var(--space-5);
-  }
-
-  .theme-family-picker {
-    grid-template-columns: repeat(3, 1fr);
   }
 }
 
@@ -1274,19 +1215,14 @@ textarea:focus {
 
 /* Section header: override .panel-header from evaluation.css */
 .settings-section .panel-header {
-  padding: var(--space-4) var(--space-6);
+  padding: var(--space-2) var(--space-5);
   margin-bottom: 0;
   border-bottom: 1px solid var(--color-border);
   background: var(--color-surface-alt);
 }
 
-.settings-section-title {
-  font-size: var(--text-xs);
-  font-weight: var(--weight-semibold);
-  color: var(--color-text-muted);
-  text-transform: uppercase;
-  letter-spacing: 0.07em;
-  margin: 0;
+.settings-section .panel-header .term-section-label {
+  padding: 0;
 }
 
 .settings-section-description {
@@ -1299,8 +1235,8 @@ textarea:focus {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
-  gap: var(--space-3);
-  padding: var(--space-5) var(--space-6);
+  gap: var(--space-2);
+  padding: var(--space-4) var(--space-5);
   border-bottom: 1px solid var(--color-border);
 }
 
@@ -1447,8 +1383,57 @@ textarea:focus {
 
 .settings-description {
   font-size: var(--text-xs);
-  color: var(--color-text-muted);
+  color: var(--color-text-subtle, var(--color-text-muted));
   line-height: 1.5;
+}
+
+/* ── Settings: terminal-aligned pill buttons ── */
+
+.settings-pill {
+  appearance: none;
+  font-family: var(--term-font-mono);
+  font-size: 12px;
+  line-height: 1;
+  padding: 8px 12px;
+  background: transparent;
+  color: var(--color-text-muted);
+  border: 1px solid var(--color-border);
+  border-radius: var(--term-radius);
+  cursor: pointer;
+  transition: background-color 0.12s, color 0.12s, border-color 0.12s;
+  white-space: nowrap;
+}
+
+.settings-pill:hover {
+  color: var(--color-text);
+  border-color: var(--color-text-muted);
+}
+
+.settings-pill--active {
+  background: var(--color-accent-soft);
+  color: var(--color-accent);
+  border-color: var(--color-accent);
+}
+
+.settings-pill--active:hover {
+  background: var(--color-accent-soft);
+  border-color: var(--color-accent);
+  color: var(--color-accent);
+}
+
+.settings-pill[disabled],
+.settings-pill[disabled]:hover {
+  cursor: not-allowed;
+  opacity: 0.5;
+  color: var(--color-text-muted);
+  border-color: var(--color-border);
+  background: transparent;
+}
+
+.settings-pill-group {
+  display: inline-flex;
+  flex-wrap: wrap;
+  gap: 6px;
 }
 
 .settings-description code {
@@ -1598,88 +1583,6 @@ textarea:focus {
 
 /* ─── Theme toggle ─── */
 
-.theme-toggle {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 8px;
-}
-
-.theme-toggle-btn {
-  padding: 10px 14px;
-  border: 2px solid var(--color-border);
-  border-radius: 8px;
-  background: var(--color-surface);
-  color: var(--color-text-muted);
-  font-family: var(--font-sans);
-  font-size: var(--text-sm);
-  font-weight: var(--weight-medium);
-  cursor: pointer;
-  transition: background 120ms ease, color 120ms ease, border-color 120ms ease;
-  white-space: nowrap;
-}
-
-.theme-toggle-btn:hover {
-  border-color: var(--color-text-muted);
-}
-
-.theme-toggle-btn.active {
-  border-color: var(--color-accent);
-  background: var(--color-accent-soft);
-  color: var(--color-accent);
-  font-weight: var(--weight-semibold);
-}
-
-.theme-family-picker {
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  gap: 8px;
-}
-
-.theme-family-card {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 10px 14px;
-  border-radius: 8px;
-  border: 2px solid var(--color-border);
-  background: var(--color-surface);
-  cursor: pointer;
-  transition: border-color 0.15s;
-  text-align: center;
-}
-
-.theme-family-card:hover {
-  border-color: var(--color-text-muted);
-}
-
-.theme-family-card.active {
-  border-color: var(--color-accent);
-  background: var(--color-accent-soft);
-}
-
-.theme-family-swatches {
-  display: flex;
-  gap: 4px;
-}
-
-.theme-family-swatch {
-  width: 16px;
-  height: 16px;
-  border-radius: 50%;
-  border: 1px solid rgba(128, 128, 128, 0.2);
-}
-
-.theme-family-label {
-  font-size: var(--text-sm);
-  font-weight: var(--weight-medium);
-  color: var(--color-text-muted);
-}
-
-.theme-family-card.active .theme-family-label {
-  color: var(--color-accent);
-  font-weight: var(--weight-semibold);
-}
-
 /* Responsive breakpoints (app shell / layout) */
 
 @media (max-width: 900px) {
@@ -1738,42 +1641,6 @@ textarea:focus {
 
 /* ── Provider Tabs ───────────────────────────────────────────────────── */
 
-.provider-tab-bar {
-  display: flex;
-  gap: 8px;
-  overflow-x: auto;
-  flex-wrap: nowrap;
-  scrollbar-width: thin;
-}
-
-.provider-tab {
-  padding: 10px 14px;
-  font-size: var(--text-sm);
-  font-weight: var(--weight-medium);
-  font-family: var(--font-sans);
-  color: var(--color-text-muted);
-  background: var(--color-surface);
-  border: 2px solid var(--color-border);
-  border-radius: 8px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  gap: var(--space-2);
-  transition: background 120ms ease, color 120ms ease, border-color 120ms ease;
-  flex-shrink: 0;
-}
-
-.provider-tab--active {
-  border-color: var(--color-accent);
-  background: var(--color-accent-soft);
-  color: var(--color-accent);
-  font-weight: var(--weight-semibold);
-}
-
-.provider-tab:hover:not(.provider-tab--active) {
-  border-color: var(--color-text-muted);
-}
-
 .provider-tab-content {
   padding: 0;
 }
@@ -1793,10 +1660,10 @@ textarea:focus {
   display: flex;
   align-items: center;
   gap: var(--space-2);
-  padding: var(--space-3) var(--space-4);
-  margin: var(--space-4);
+  padding: var(--space-2) var(--space-3);
+  margin: var(--space-3) var(--space-5);
   border-radius: var(--radius-md);
-  font-size: var(--text-sm);
+  font-size: var(--text-xs);
 }
 
 .server-status--online {
